Refactor content handlers, first step * Move most of logic into ApplicationRepository * Remove unnneded base class
